requirements.txt
asgiref==3.4.1
click==8.0.3
cmake==3.22.1
colorama==0.4.4
Django==3.2.10
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0

importlib-metadata==4.9.0
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.6.0
